requirements.txt
# PatentScout dependencies
streamlit>=1.28.0
google-cloud-bigquery>=3.12.0
google-cloud-bigquery-storage>=2.13.0
google-genai>=1.0.0
google-cloud-aiplatform>=1.38.0
sentence-transformers>=2.2.0
scikit-learn>=1.3.0
pandas>=2.0.0
plotly>=5.18.0
matplotlib>=3.7.0
reportlab>=4.0.0
Pillow>=10.0.0
pdfplumber>=0.10.0
PyMuPDF>=1.23.0
numpy>=1.24.0
python-dotenv>=1.0.0
kaleido>=0.2.1
db-dtypes>=1.1.0
pyarrow>=12.0.0
